Andrew Smith 6/29/2011 4:04PM
This version of the Iron Golem has all updates previous including the panels for dangling wires in the torso and spinal construct within the torso.

The latest inclusion is the furnace upon the Iron Golem's back.  The mouth at the furnace is its own object and can move.  The exact positions and amount of pipes on the backpack is up for debate.