Andrew Smith 6/30/2011 1:20PM
This is the general concept for the power bar that sits at the bottom middle of the screen.  This will be seen at almost all times so it needs to either be pretty or at least not in the way.  The final product should probably be a 3D model that is loaded and placed when the game loads.