This is the Maintenance Bug monster, a large mechanical caterpillar.
Maintenance_Bug_08.max: Is the mesh of the monster.
Render.bmp: Is a render of the monster.

Notes: I left the tentacles upright in the middle which may help with the rigging.  I can bend them if need be.  Animations for this will be fun and I gave overlap in the chitin that'll aid with getting the scrunching and expanding of the caterpillar movement.